+++ src/Server/InverterService/Commands/Command.cs
@@ -3,6 +3,7 @@
 public interface ICommand
 {
     string CommandString { get; set; }
+    bool IsTroublesomeCmd { get; }
     void Parse(string rawResponse);
     void Start();
     void End();
@@ -11,6 +12,7 @@ public interface ICommand
 public abstract class Command<TResponseDto> : ICommand where TResponseDto : new()
 {
     public abstract string CommandString { get; set; }
+    public virtual bool IsTroublesomeCmd { get; }
     public TResponseDto Result { get; protected set; } = new();
     public bool IsComplete { get; private set; }
 
@@ -27,9 +29,9 @@ public abstract class Command<TResponseDto> : ICommand where TResponseDto : new(
     public void End()
         => IsComplete = true;
 
-    public async Task WhileProcessing(CancellationToken c)
+    public async Task WhileProcessing(CancellationToken c, int timeoutMillis = Constants.StatusPollingFrequencyMillis)
     {
-        while (!c.IsCancellationRequested && !IsComplete && DateTime.Now.Subtract(startTime).TotalMilliseconds <= Constants.StatusPollingFrequencyMillis)
+        while (!c.IsCancellationRequested && !IsComplete && DateTime.Now.Subtract(startTime).TotalMilliseconds <= timeoutMillis)
             await Task.Delay(500, c);
     }
 }