+++ src/Server/InverterService/Commands/GetChargeAmpereValues.cs
@@ -3,9 +3,12 @@ namespace InverterMon.Server.InverterService.Commands;
 internal class GetChargeAmpereValues : Command<List<string>>
 {
     public override string CommandString { get; set; } = "QMCHGCR";
+    public override bool IsTroublesomeCmd { get; } = true;
 
     public GetChargeAmpereValues(bool getUtilityValues)
     {
+        Result.AddRange(new[] { "000" });
+
         if (getUtilityValues)
             CommandString = "QMUCHGCR";
     }
@@ -13,14 +16,14 @@ internal class GetChargeAmpereValues : Command<List<string>>
     public override void Parse(string responseFromInverter)
     {
         if (responseFromInverter.StartsWith("(NAK"))
-        {
-            Result.Add("000");
             return;
-        }
 
         var parts = responseFromInverter[1..]
             .Split(' ', StringSplitOptions.RemoveEmptyEntries)
-            .Select(x => x[0..3]);
+            .Select(x => x[..3]);
+
+        if (parts.Any())
+            Result.Clear(); //remove default values
 
         Result.AddRange(parts);
     }