+++ src/Server/InverterService/Commands/SetSetting.cs
@@ -3,6 +3,7 @@ namespace InverterMon.Server.InverterService.Commands;
 internal class SetSetting : Command<bool>
 {
     public override string CommandString { get; set; }
+    public override bool IsTroublesomeCmd { get; } = true;
 
     public SetSetting(string settingName, string settingValue)
     {